Move CopyScheduler to smart-hadoop-support
